+++ tests/integration/gov/abci_test.go
@@ -24,7 +24,7 @@ import (
 
 func TestUnregisteredProposal_InactiveProposalFails(t *testing.T) {
 	suite := createTestSuite(t)
-	ctx := suite.App.BaseApp.NewContext(false)
+	ctx := suite.app.BaseApp.NewContext(false)
 	addrs := simtestutil.AddTestAddrs(suite.BankKeeper, suite.StakingKeeper, ctx, 10, valTokens)
 
 	// manually set proposal in store
@@ -50,7 +50,7 @@ func TestUnregisteredProposal_InactiveProposalFails(t *testing.T) {
 
 func TestUnregisteredProposal_ActiveProposalFails(t *testing.T) {
 	suite := createTestSuite(t)
-	ctx := suite.App.BaseApp.NewContext(false)
+	ctx := suite.app.BaseApp.NewContext(false)
 	addrs := simtestutil.AddTestAddrs(suite.BankKeeper, suite.StakingKeeper, ctx, 10, valTokens)
 
 	// manually set proposal in store
@@ -79,7 +79,7 @@ func TestUnregisteredProposal_ActiveProposalFails(t *testing.T) {
 
 func TestTickExpiredDepositPeriod(t *testing.T) {
 	suite := createTestSuite(t)
-	app := suite.App
+	app := suite.app
 	ctx := app.BaseApp.NewContext(false)
 	addrs := simtestutil.AddTestAddrs(suite.BankKeeper, suite.StakingKeeper, ctx, 10, valTokens)
 
@@ -115,7 +115,7 @@ func TestTickExpiredDepositPeriod(t *testing.T) {
 
 func TestTickMultipleExpiredDepositPeriod(t *testing.T) {
 	suite := createTestSuite(t)
-	app := suite.App
+	app := suite.app
 	ctx := app.BaseApp.NewContext(false)
 	addrs := simtestutil.AddTestAddrs(suite.BankKeeper, suite.StakingKeeper, ctx, 10, valTokens)
 	govMsgSvr := keeper.NewMsgServerImpl(suite.GovKeeper)
@@ -169,7 +169,7 @@ func TestTickMultipleExpiredDepositPeriod(t *testing.T) {
 
 func TestTickPassedDepositPeriod(t *testing.T) {
 	suite := createTestSuite(t)
-	app := suite.App
+	app := suite.app
 	ctx := app.BaseApp.NewContext(false)
 	addrs := simtestutil.AddTestAddrs(suite.BankKeeper, suite.StakingKeeper, ctx, 10, valTokens)
 	govMsgSvr := keeper.NewMsgServerImpl(suite.GovKeeper)
@@ -204,7 +204,7 @@ func TestTickPassedDepositPeriod(t *testing.T) {
 
 func TestProposalDepositRefundFailEndBlocker(t *testing.T) {
 	suite := createTestSuite(t)
-	app := suite.App
+	app := suite.app
 	ctx := app.BaseApp.NewContext(false)
 	addrs := simtestutil.AddTestAddrs(suite.BankKeeper, suite.StakingKeeper, ctx, 10, valTokens)
 	govMsgSvr := keeper.NewMsgServerImpl(suite.GovKeeper)
@@ -272,7 +272,7 @@ func TestTickPassedVotingPeriod(t *testing.T) {
 	for _, tc := range testcases {
 		t.Run(tc.name, func(t *testing.T) {
 			suite := createTestSuite(t)
-			app := suite.App
+			app := suite.app
 			ctx := app.BaseApp.NewContext(false)
 			depositMultiplier := getDepositMultiplier(tc.proposalType)
 			addrs := simtestutil.AddTestAddrs(suite.BankKeeper, suite.StakingKeeper, ctx, 10, valTokens.Mul(math.NewInt(depositMultiplier)))
@@ -349,7 +349,7 @@ func TestProposalPassedEndblocker(t *testing.T) {
 	for _, tc := range testcases {
 		t.Run(tc.name, func(t *testing.T) {
 			suite := createTestSuite(t)
-			app := suite.App
+			app := suite.app
 			ctx := app.BaseApp.NewContext(false)
 			depositMultiplier := getDepositMultiplier(tc.proposalType)
 			addrs := simtestutil.AddTestAddrs(suite.BankKeeper, suite.StakingKeeper, ctx, 10, valTokens.Mul(math.NewInt(depositMultiplier)))
@@ -404,7 +404,7 @@ func TestProposalPassedEndblocker(t *testing.T) {
 
 func TestEndBlockerProposalHandlerFailed(t *testing.T) {
 	suite := createTestSuite(t)
-	app := suite.App
+	app := suite.app
 	ctx := app.BaseApp.NewContext(false)
 	addrs := simtestutil.AddTestAddrs(suite.BankKeeper, suite.StakingKeeper, ctx, 1, valTokens)
 
@@ -485,7 +485,7 @@ func TestExpeditedProposal_PassAndConversionToRegular(t *testing.T) {
 	for _, tc := range testcases {
 		t.Run(tc.name, func(t *testing.T) {
 			suite := createTestSuite(t)
-			app := suite.App
+			app := suite.app
 			ctx := app.BaseApp.NewContext(false)
 			depositMultiplier := getDepositMultiplier(v1.ProposalType_PROPOSAL_TYPE_EXPEDITED)
 			addrs := simtestutil.AddTestAddrs(suite.BankKeeper, suite.StakingKeeper, ctx, 3, valTokens.Mul(math.NewInt(depositMultiplier)))
