+++ tests/integration/gov/genesis_test.go
@@ -6,11 +6,13 @@ import (
 
 	abci "github.com/cometbft/cometbft/abci/types"
 	dbm "github.com/cosmos/cosmos-db"
+	"github.com/stretchr/testify/require"
 	"gotest.tools/v3/assert"
 
 	"cosmossdk.io/core/header"
 	"cosmossdk.io/depinject"
 	"cosmossdk.io/log"
+	sdkmath "cosmossdk.io/math"
 	_ "cosmossdk.io/x/auth"
 	authkeeper "cosmossdk.io/x/auth/keeper"
 	authtypes "cosmossdk.io/x/auth/types"
@@ -195,4 +197,30 @@ func clearDB(t *testing.T, db *dbm.MemDB) {
 	for _, k := range keys {
 		assert.NilError(t, db.Delete(k))
 	}
+}
+
+func TestImportExportQueues_ErrorUnconsistentState(t *testing.T) {
+	suite := createTestSuite(t)
+	app := suite.app
+	ctx := app.BaseApp.NewContext(false)
+	require.Panics(t, func() {
+		gov.InitGenesis(ctx, suite.AccountKeeper, suite.BankKeeper, suite.GovKeeper, &v1.GenesisState{
+			Deposits: v1.Deposits{
+				{
+					ProposalId: 1234,
+					Depositor:  "me",
+					Amount: sdk.Coins{
+						sdk.NewCoin(
+							"stake",
+							sdkmath.NewInt(1234),
+						),
+					},
+				},
+			},
+		})
+	})
+	gov.InitGenesis(ctx, suite.AccountKeeper, suite.BankKeeper, suite.GovKeeper, v1.DefaultGenesisState())
+	genState, err := gov.ExportGenesis(ctx, suite.GovKeeper)
+	require.NoError(t, err)
+	require.Equal(t, genState, v1.DefaultGenesisState())
 }
