--- x/gov/genesis_test.go
@@ -1,39 +1,0 @@
-package gov_test
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/require"
-
-	sdkmath "cosmossdk.io/math"
-	"cosmossdk.io/x/gov"
-	v1 "cosmossdk.io/x/gov/types/v1"
-
-	sdk "github.com/cosmos/cosmos-sdk/types"
-)
-
-func TestImportExportQueues_ErrorUnconsistentState(t *testing.T) {
-	suite := createTestSuite(t)
-	app := suite.App
-	ctx := app.BaseApp.NewContext(false)
-	require.Panics(t, func() {
-		gov.InitGenesis(ctx, suite.AccountKeeper, suite.BankKeeper, suite.GovKeeper, &v1.GenesisState{
-			Deposits: v1.Deposits{
-				{
-					ProposalId: 1234,
-					Depositor:  "me",
-					Amount: sdk.Coins{
-						sdk.NewCoin(
-							"stake",
-							sdkmath.NewInt(1234),
-						),
-					},
-				},
-			},
-		})
-	})
-	gov.InitGenesis(ctx, suite.AccountKeeper, suite.BankKeeper, suite.GovKeeper, v1.DefaultGenesisState())
-	genState, err := gov.ExportGenesis(ctx, suite.GovKeeper)
-	require.NoError(t, err)
-	require.Equal(t, genState, v1.DefaultGenesisState())
-}
